fixed signed/unsigned mismatch in the comparsion
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@73805 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/MachineImpl.cpp
@@ -5076,7 +5076,7 @@ HRESULT Machine::setExtraData(const com::Utf8Str &aKey, const com::Utf8Str &aVal
      * they are rejected unless the key should be deleted. */
     if (!aValue.isEmpty())
     {
-        for (int i = 0; i < aKey.length(); ++i)
+        for (size_t i = 0; i < aKey.length(); ++i)
         {
             char ch = aKey[i];
             if (!RTLocCIsPrint(ch))
